hmap: switch size and mask to size_t
Size is already a size_t in the pre-assignment logic, and mask needs to
be changed as well because it is directly derived from the size itself.

+++ libnopegl/src/hmap.c
@@ -35,8 +35,8 @@ struct bucket {
 
 struct hmap {
     struct bucket *buckets;
-    int size;
-    uint32_t mask;
+    size_t size;
+    size_t mask;
     int count; // total number of entries
     user_free_func_type user_free_func;
     void *user_arg;
